Fixed first IP64 port mapping not being randomized.
The ports for the IP64 module are randomized since commit
ad744c621866ec73b592707c6210ccc7ccbfa387, for security reasons. The very first
port that the module would allocate is not randomized, though, and will always
be 10000.

In the process I noticed that the ip64_addrmap_init() isn't called anywhere. I
propose calling it in the init function of the ip64 module, which in turn is
called in several border-router examples.